Abstract CppCodeGenerator to support C++/WinRT in the future
