+++ src/VSExtension/ReswPlusVSIX/ReswPlusVSIX.csproj
@@ -59,13 +59,14 @@
     <Compile Include="ClassGenerator\Models\StronglyTypedClass.cs" />
     <Compile Include="ContextMenu.cs" />
     <Compile Include="Languages\CodeStringBuilder.cs" />
+    <Compile Include="Languages\CppCXCodeGenerator.cs" />
     <Compile Include="Languages\CsharpCodeGenerator.cs" />
     <Compile Include="Languages\DotNetGeneratorBase.cs" />
     <Compile Include="Languages\VBCodeGenerator.cs" />
     <Compile Include="Languages\ICodeGenerator.cs" />
     <Compile Include="SingleFileGenerators\ReswPlusGeneratorBase.cs" />
     <Compile Include="SingleFileGenerators\ReswPlusAdvancedGenerator.cs" />
-    <Compile Include="Languages\CppCodeGenerator.cs" />
+    <Compile Include="Languages\CppCodeGeneratorBase.cs" />
     <Compile Include="Resw\ParameterType.cs" />
     <Compile Include="Resw\ReswParser.cs" />
     <Compile Include="Resw\ReswTagTyped.cs" />
